requirements.txt
aws-cdk-lib==2.45.0
aws-cdk.aws-lambda-python-alpha==2.45.0a0
aws-lambda-powertools==1.31.1
constructs>=10.0.0,<11.0.0
emoji==2.1.0
aws_solutions_constructs.aws_eventbridge_lambda
boto3==1.26.17